Introduce gammapy.cube sub-package
