+++ gammapy/cube/__init__.py
@@ -1,0 +1,5 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+
+from .core import *
+from .spectral_cube import *
+from .exposure import *
