+++ gammapy/cube/exposure.py
@@ -2,6 +2,8 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 import numpy as np
 from astropy.coordinates import SkyCoord, Angle
+from .spectral_cube import SpectralCube
+
 
 __all__ = [
     'exposure_cube'
@@ -38,8 +40,7 @@ def exposure_cube(pointing,
     # TODO: figure out if `gammapy.irf` is a good location for
     # exposure computation functionality, or if this should be
     # moved to `gammapy.data` or `gammapy.spectrum` or ...
-    from ..data import SpectralCube
-
+    
     ny, nx = ref_cube.data.shape[1:]
     xx, yy = np.meshgrid(np.arange(nx), np.arange(ny))
     lon, lat, en = ref_cube.pix2world(xx, yy, 0)
