+++ gammapy/cube/tests/test_exposure.py
