+++ gammapy/cube/tests/test_spectral_cube.py
